pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "malphas"
version = "0.1.0"
description = "Privacy-first P2P messenger with onion routing"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=42.0",
    "fastapi>=0.111",
    "uvicorn>=0.29",
    "websockets>=12.0",
    "zeroconf>=0.131",
    "pydantic>=2.0",
    "stem>=1.8",
    "argon2-cffi>=23.0",
]

[project.scripts]
malphas = "malphas.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/malphas"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
